Consolidate ConfigurationParser into WorkflowRunner
